test_suite_executor/pyproject.toml
[project]
name = "test-suite-executor"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "vikentiy.smetanin", email = "vikentiy.smetanin@axxonsoft.dev" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "gitpython>=3.1.44",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "test-suite[pytest]",
    "uvicorn[standard]>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
test-suite = { path = "../test_suite", editable = true }
